+++ CHANGELOG.md
@@ -21,10 +21,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changelog Formatting by [@vrdons](https://github.com/vrdons) ([507277f](https://github.com/vrdons/module-template/commit/507277fae484d98bcd828b29625eeebbc9e7b25e))
 - Markdown  config by [@vrdons](https://github.com/vrdons) ([507277f](https://github.com/vrdons/module-template/commit/507277fae484d98bcd828b29625eeebbc9e7b25e))
 - Github release unknown owner by [@vrdons](https://github.com/vrdons) ([1a0a7ec](https://github.com/vrdons/module-template/commit/1a0a7ec20d185bf0216559c13361b2e403fb37b9))
+- Github and npm release by [@vrdons](https://github.com/vrdons) ([a2d040f](https://github.com/vrdons/module-template/commit/a2d040fb899a409eed15ae96a434692079162f3d))
 
 ### 💄 Styles
 
 - Update script code style by [@vrdons](https://github.com/vrdons) ([9fb186d](https://github.com/vrdons/module-template/commit/9fb186dd3cf9aef61877771afc91aac122a060e3))
+
+### ✅ Tests
+
+- Npm version check by [@vrdons](https://github.com/vrdons) ([a2d040f](https://github.com/vrdons/module-template/commit/a2d040fb899a409eed15ae96a434692079162f3d))
 
 ### 📦 Dependency Updates